Rename BenchmarkRunnerOptions to BenchmarkCommand
