+++ Sources/Benchmark/BenchmarkRunner.swift
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import ArgumentParser
-import Foundation
-
 public struct BenchmarkRunner {
     let suites: [BenchmarkSuite]
     let reporter: BenchmarkReporter
@@ -25,16 +22,16 @@ public struct BenchmarkRunner {
         self.reporter = reporter
     }
 
-    mutating func run(options: BenchmarkRunnerOptions) {
+    mutating func run(command: BenchmarkCommand) {
         for suite in suites {
-            run(suite: suite, options: options)
+            run(suite: suite, command: command)
         }
         reporter.report(results: results)
     }
 
-    mutating func run(suite: BenchmarkSuite, options: BenchmarkRunnerOptions) {
+    mutating func run(suite: BenchmarkSuite, command: BenchmarkCommand) {
         for benchmark in suite.benchmarks {
-            if !options.matches(suiteName: suite.name, benchmarkName: benchmark.name) { continue }
+            if !command.matches(suiteName: suite.name, benchmarkName: benchmark.name) { continue }
             run(benchmark: benchmark, suite: suite)
         }
     }
@@ -62,65 +59,5 @@ public struct BenchmarkRunner {
             suiteName: suite.name,
             measurements: measurements)
         results.append(result)
-    }
-}
-
-/// Allows dynamic configuration of the benchmark execution.
-internal struct BenchmarkRunnerOptions: ParsableCommand {
-    @Option(
-        help: "Run only benchmarks whose names match the regular expression.",
-        transform: BenchmarkFilter.init)
-    var filter: BenchmarkFilter?
-
-    @Flag(help: "Overrides check to verify optimized build.")
-    var allowDebugBuild: Bool
-
-    mutating func validate() throws {
-        var isDebug = false
-        assert(
-            {
-                isDebug = true
-                return true
-            }())
-        if isDebug && !allowDebugBuild {
-            throw ValidationError(debugBuildErrorMessage)
-        }
-    }
-
-    var debugBuildErrorMessage: String {
-        """
-        Please build with optimizations enabled (`-c release` if using SwiftPM,
-        `-c opt` if using bazel, or `-O` if using swiftc directly). If you would really
-        like to run the benchmark without optimizations, pass the `--allow-debug-build`
-        flag.
-        """
-    }
-}
-
-extension BenchmarkRunnerOptions {
-    func matches(suiteName: String, benchmarkName: String) -> Bool {
-        guard let filter = filter else { return true }
-        return filter.matches(suiteName: suiteName, benchmarkName: benchmarkName)
-    }
-
-    init(filter: String) throws {
-        self.filter = try BenchmarkFilter(filter)
-        self.allowDebugBuild = false
-    }
-}
-
-internal struct BenchmarkFilter {
-    let underlying: NSRegularExpression
-
-    init(_ regularExpression: String) throws {
-        underlying = try NSRegularExpression(
-            pattern: regularExpression,
-            options: [.caseInsensitive, .anchorsMatchLines])
-    }
-
-    func matches(suiteName: String, benchmarkName: String) -> Bool {
-        let str = "\(suiteName)/\(benchmarkName)"
-        let range = NSRange(location: 0, length: str.utf16.count)
-        return underlying.firstMatch(in: str, range: range) != nil
     }
 }
